services/rule_based_event_detection/requirements.txt
paho-mqtt==1.6.1
rdflib==6.3.2
pyshacl==0.25.0